add example and readme
