+++ src/TemplateProcessor.ts
@@ -27,7 +27,7 @@ import { LOG_LEVELS } from "./ConsoleLogger.js";
 import {TimerManager} from "./TimerManager.js";
 import { stringifyTemplateJSON } from './utils/stringify.js';
 import {debounce} from "./utils/debounce.js"
-import {rateLimit} from "./utils/ratelimit.js"
+import {rateLimit} from "./utils/rateLimit.js"
 
 type MetaInfoMap = Record<JsonPointerString, MetaInfo[]>;
 export type StatedError = {
